LayoutTests/http/tests/security/xss-DENIED-javascript-with-spaces-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/innocent-victim-with-notify.html from frame with URL http://127.0.0.1:8000/security/xss-DENIED-javascript-with-spaces.html. Domains, protocols and ports must match.


There should be no alert displayed.
